requirements.txt
streamlit>=1.39
duckdb>=1.1.1
polars>=1.6.0
plotly>=5.24.1
pyyaml>=6.0.1
pandas>=2.2.2
pyarrow>=17.0.0
